refactor: remove CardModal and cardsApi.js files to streamline codebase
